requirements.txt
livekit-agents[openai,anthropic,google,azure,deepgram,elevenlabs,cartesia,assemblyai,silero,turn-detector,mcp,bey]~=1.3
livekit-plugins-noise-cancellation~=0.2
cryptography>=41.0.0
python-dotenv~=1.0.0
Flask>=3.0.0
flask-cors>=4.0.0
SQLAlchemy>=2.0.35
psycopg[binary]>=3.2.1
alembic>=1.13.2
soundfile~=0.12.1
numpy>=1.26.0
scipy>=1.11.0
requests>=2.32.5
aiohttp>=3.10.0
websockets>=13.0
boto3>=1.34.0
pydub~=0.25.1
twilio>=9.0.0
mem0ai>=0.1.0
ormsgpack>=1.5.0
